Update for model card
