+++ scripts/common/constants.js
@@ -10,3 +10,4 @@ export const PROD = {
   hostName: 'https://publish-p132653-e1294163.adobeaemcloud.com',
 };
 export const disclaimerGQlEndpoint = '/graphql/execute.json/metafox/getDisclaimerCFModel;disclaimerCfPath=';
+export const wdhModelDetailsEndpoint = '/WDH_API/Models/ModelDetails/';
